Add flatmap tests
